requirements.txt
langgraph
llama-index
groq
sympy
scipy
numpy
python-dotenv
streamlit
plotly
